+++ tests/fixtures/uses_strict_mode_partially_recorded/packs/foo/app/services/foo.rb
@@ -1,0 +1,9 @@
+module Foo
+  def calls_bar_without_stated_dependency
+    Bar
+  end
+
+  def calls_baz_without_stated_dependency
+    Baz
+  end
+end